Handling x variables with factor = 0.
You don't have to specify if the factor of an x is 0
like before (e.g. 0x1 + x2 + 0x3 + x4). If you did't inputted
the problem like that and just emitted the x variables
with factor = 0  the output matrixes would be incorrect.
This bug is fixed. You can now emit x variables with factor = 0
and the script will handle it.